src/third_party/blink/web_tests/external/wpt/dom/events/Event-dispatch-handlers-changed-expected.txt

This is a testharness.js-based test.
FAIL  Dispatch additional events inside an event listener  assert_array_equals: actual_targets lengths differ, expected array [object "[object Window]", Document node with 2 children, Element node <html><head><meta charset="utf-8">
<title> Dispatch addit..., Element node <body><div id="log"></div>

<table id="table" border="1" ..., Element node <table id="table" border="1" style="display: none">
    <..., Element node <tbody id="table-body">
    <tr id="table-row">
        <..., Element node <tr id="parent">
        <td id="target">Over the river, ..., Element node <td id="target">Over the river, Charlie</td>, [...], [...], [...], [...], [...], [...], [...], [...], [...]] length 17, got [object "[object Window]", Document node with 2 children, Element node <html><head><meta charset="utf-8">
<title> Dispatch addit..., Element node <body><div id="log"></div>

<table id="table" border="1" ..., Element node <table id="table" border="1" style="display: none">
    <..., Element node <tbody id="table-body">
    <tr id="table-row">
        <..., Element node <tr id="parent">
        <td id="target">Over the river, ..., Element node <td id="target">Over the river, Charlie</td>, [...], [...], [...], [...], [...], [...], [...], [...]] length 16
Harness: the test ran to completion.
